+++ ScreenShare/Device.swift
@@ -30,7 +30,7 @@ class Device: NSObject, NSCoding {
     }
     
     
-    static let ArchivePath = NSHomeDirectory().stringByAppendingString("/devices")
+    static let ArchivePath = NSHomeDirectory().appending("/devices")
 
     convenience init?(fromDevice device: AVCaptureDevice) {
         self.init(name: device.localizedName, uid: device.uniqueID, portraitRect:NSRect(), landscapeRect:NSRect())
@@ -63,20 +63,20 @@ class Device: NSObject, NSCoding {
 
     
     //MARK: NSCoding
-    func encodeWithCoder(aCoder: NSCoder) {
-        aCoder.encodeObject(name, forKey: PropertyKey.nameKey)
-        aCoder.encodeObject(uid, forKey: PropertyKey.uidKey)
+    func encode(with aCoder: NSCoder) {
+        aCoder.encode(name, forKey: PropertyKey.nameKey)
+        aCoder.encode(uid, forKey: PropertyKey.uidKey)
         
-        aCoder.encodeObject(NSStringFromRect(portraitRect), forKey: PropertyKey.portraitRectKey)
-        aCoder.encodeObject(NSStringFromRect(landscapeRect), forKey: PropertyKey.landscapeRectKey)
+        aCoder.encode(NSStringFromRect(portraitRect), forKey: PropertyKey.portraitRectKey)
+        aCoder.encode(NSStringFromRect(landscapeRect), forKey: PropertyKey.landscapeRectKey)
         
     }
     
     required convenience init?(coder aDecoder: NSCoder) {
-        guard let name = aDecoder.decodeObjectForKey(PropertyKey.nameKey) as? String,
-            let uid = aDecoder.decodeObjectForKey(PropertyKey.uidKey) as? String,
-            let portraitRect = aDecoder.decodeObjectForKey(PropertyKey.portraitRectKey) as? String,
-            let landscapeRect = aDecoder.decodeObjectForKey(PropertyKey.landscapeRectKey) as? String else {
+        guard let name = aDecoder.decodeObject(forKey: PropertyKey.nameKey) as? String,
+            let uid = aDecoder.decodeObject(forKey: PropertyKey.uidKey) as? String,
+            let portraitRect = aDecoder.decodeObject(forKey: PropertyKey.portraitRectKey) as? String,
+            let landscapeRect = aDecoder.decodeObject(forKey: PropertyKey.landscapeRectKey) as? String else {
                 return nil
         }